made second function return long long

+++ recursion/main.h
@@ -7,7 +7,7 @@ void _print_rev_recursion(char *s);
 int _strlen_recursion(char *s);
 int factorial(int n);
 int _pow_recursion(int x, int y);
-long int _sqrt_part_two(int n, int m);
+long long _sqrt_part_two(int n, int m);
 int _sqrt_recursion(int n);
 int is_prime_number(int n);
